The `schema` is correctly respected now for the the `cursors` table
We are also heavily refactored how operations keeps schema details like a table and columns. We now compute once for the full schema the requires escaped version of all tables and columns.

This results in greatly educed allocations and escaping performed which should increase ingestion speed, this will be more visible for Substreams where a lot of entities and columns are processed.

Fixes #21